TEST: Migrate TicketTailorProvider tests from Feature to Unit directory and enhance test coverage
